refactor: :truck: Update project file structure
Add feature based component structure, remove unused code
