+++ components/features/scanning/QRCodeScanner.tsx
