+++ components/shared/SafeAreaContent.tsx
